tcpip: give enough time to the instrument to send a timeout message
The VXI-11 protocol defines that if an instrument cannot answer the server will send a message saying so. We must wait long enough for this message to arrive but still timeout at the socket level in case something else went wrong (cable disconnected,...) in which case it is an io error and not a timeout.