fix crash on Windows when opening multiple devices
For some reason, PyUSB on Windows is unable to read the serial_number
attribute from a device that's already open in another pyusb instance,
and crashes with a cryptic "device has no langid" exception when it's accessed.

When a USB resource is opened, PyVISA-py tries to find the device by serial
number. If there's already another open device, its serial_number
attribute will be accessed in this search, triggering the langid
exception.

With this fix we will silently ignore a device when accessing a search
attribute on it raises an exception.