readme: deduplicate requirement version specification
Specifying the dependency versions in the README file creates an extra
maintenance burden not worth the effort. In fact, it has been left
outdated more than once leading to misinformation [1, 2, 3]. Instead,
provide the rough information in the documentation and point users to
the source of truth for details.

Python 3 is still listed explicitly since it won't be bumped every now
and then, and provides a useful compatibility information for users.
Unfortunately, PyVISA cannot be dealt with in a similar vein.

[1] dd8e9041bb9d ("switch to a PEP517 install")
[2] 8a0753b4e23d ("bump Python requirement to 3.8 and PyVISA requirement to 1.13")
[3] 29d3048d23e2 ("Drop support for Python 3.8 and 3.9 and add support for Python 3.13")